Renamed UdfDebugger to UdfOutputLogger
Renamed file script_output_redirect.py to consumer.py
