+++ exasol/pytest_slc/udf_debug/__init__.py
@@ -1,9 +1,13 @@
 from exasol.pytest_slc.udf_debug.ip_address import IpAddress
 from exasol.pytest_slc.udf_debug.messages import wait_for_messages
-from exasol.pytest_slc.udf_debug.udf_debugger import UdfDebugger
+from exasol.pytest_slc.udf_debug.udf_output_logger import (
+    UdfOutputLogger,
+    alter_session_sql,
+)
 
 __all__ = [
-    "UdfDebugger",
     "IpAddress",
+    "UdfOutputLogger",
+    "alter_session_sql",
     "wait_for_messages",
 ]
