+++ exasol/pytest_slc/udf_debug/types.py
@@ -1,0 +1,9 @@
+from typing import (
+    Callable,
+    TypeAlias,
+)
+
+import pyexasol
+
+PrintFunc: TypeAlias = Callable[[str], None]
+QueryFunc: TypeAlias = Callable[[str], pyexasol.ExaStatement]
